libs/common: refactor libs/common 2
- move random function to there own pkg
