+++ blockchain/v1/peer_test.go
@@ -8,15 +8,15 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	cmn "github.com/tendermint/tendermint/libs/common"
-	"github.com/tendermint/tendermint/libs/log"
+	github.com/tendermint/tendermint/libs/rlog
+	"github.com/tendermint/tendermint/libs/rand"
 	"github.com/tendermint/tendermint/p2p"
 	"github.com/tendermint/tendermint/types"
 )
 
 func TestPeerMonitor(t *testing.T) {
 	peer := NewBpPeer(
-		p2p.ID(cmn.RandStr(12)), 10,
+		p2p.ID(rand.RandStr(12)), 10,
 		func(err error, _ p2p.ID) {},
 		nil)
 	peer.SetLogger(log.TestingLogger())
@@ -35,7 +35,7 @@ func TestPeerResetBlockResponseTimer(t *testing.T) {
 	params := &BpPeerParams{timeout: 2 * time.Millisecond}
 
 	peer := NewBpPeer(
-		p2p.ID(cmn.RandStr(12)), 10,
+		p2p.ID(rand.RandStr(12)), 10,
 		func(err error, _ p2p.ID) {
 			peerTestMtx.Lock()
 			defer peerTestMtx.Unlock()
@@ -75,7 +75,7 @@ func TestPeerRequestSent(t *testing.T) {
 	params := &BpPeerParams{timeout: 2 * time.Millisecond}
 
 	peer := NewBpPeer(
-		p2p.ID(cmn.RandStr(12)), 10,
+		p2p.ID(rand.RandStr(12)), 10,
 		func(err error, _ p2p.ID) {},
 		params)
 
@@ -94,7 +94,7 @@ func TestPeerRequestSent(t *testing.T) {
 
 func TestPeerGetAndRemoveBlock(t *testing.T) {
 	peer := NewBpPeer(
-		p2p.ID(cmn.RandStr(12)), 100,
+		p2p.ID(rand.RandStr(12)), 100,
 		func(err error, _ p2p.ID) {},
 		nil)
 
@@ -142,7 +142,7 @@ func TestPeerGetAndRemoveBlock(t *testing.T) {
 
 func TestPeerAddBlock(t *testing.T) {
 	peer := NewBpPeer(
-		p2p.ID(cmn.RandStr(12)), 100,
+		p2p.ID(rand.RandStr(12)), 100,
 		func(err error, _ p2p.ID) {},
 		nil)
 
@@ -189,7 +189,7 @@ func TestPeerOnErrFuncCalledDueToExpiration(t *testing.T) {
 	)
 
 	peer := NewBpPeer(
-		p2p.ID(cmn.RandStr(12)), 10,
+		p2p.ID(rand.RandStr(12)), 10,
 		func(err error, _ p2p.ID) {
 			peerTestMtx.Lock()
 			defer peerTestMtx.Unlock()
@@ -215,7 +215,7 @@ func TestPeerCheckRate(t *testing.T) {
 		minRecvRate: int64(100), // 100 bytes/sec exponential moving average
 	}
 	peer := NewBpPeer(
-		p2p.ID(cmn.RandStr(12)), 10,
+		p2p.ID(rand.RandStr(12)), 10,
 		func(err error, _ p2p.ID) {},
 		params)
 	peer.SetLogger(log.TestingLogger())
@@ -249,7 +249,7 @@ func TestPeerCleanup(t *testing.T) {
 	params := &BpPeerParams{timeout: 2 * time.Millisecond}
 
 	peer := NewBpPeer(
-		p2p.ID(cmn.RandStr(12)), 10,
+		p2p.ID(rand.RandStr(12)), 10,
 		func(err error, _ p2p.ID) {},
 		params)
 	peer.SetLogger(log.TestingLogger())
